+++ intervention-results/intervention-results.ts
@@ -4,7 +4,7 @@ import './supply-agreement/supply-agreement';
 import './results-structure/results-structure';
 import './effective-efficient-programme-mgmt/effective-efficient-programme-mgmt';
 import {getStore} from '../utils/redux-store-access';
-import {commentsData} from '../common/comments-components/comments.reducer';
+import {commentsData} from '../common/components/comments/comments.reducer';
 
 getStore().addReducers({
   commentsData
